Remove unused facial recognition files. Prep setup for bootstrap installation
