--- nipo/recog.py
@@ -1,104 +1,0 @@
-
-import face_recognition
-import cv2
-
-# This is a super simple (but slow) example of running face recognition on live video from your webcam.
-# There's a second example that's a little more complicated but runs faster.
-
-# PLEASE NOTE: This example requires OpenCV (the `cv2` library) to be installed only to read from your webcam.
-# OpenCV is *not* required to use the face_recognition library. It's only required if you want to run this
-# specific demo. If you have trouble installing it, try any of the other demos that don't require it instead.
-
-#Just to reiterate, this example is simply a proof of concept, the file recognise.py will eventually be where the magic happens.
-
-# Get a reference to webcam #0 (the default one)
-video_capture = cv2.VideoCapture(0)
-
-# Load a sample picture and learn how to recognize it.
-jos_image = face_recognition.load_image_file("josiah.JPG")
-jos_face_encoding = face_recognition.face_encodings(jos_image)[0]
-
-# Create arrays of known face encodings and their names
-known_face_encodings = [
-    jos_face_encoding]
-known_face_names = [
-    "Josiah Nyangaga"
-]
-
-# Initialize some variables
-face_locations = []
-face_encodings = []
-face_names = []
-process_this_frame = True
-
-while True:
-    # Grab a single frame of video
-    ret, frame = video_capture.read()
-
-    # Resize frame of video to 1/4 size for faster face recognition processing
-    small_frame = cv2.resize(frame, (0, 0), fx=0.25, fy=0.25)
-
-    # Convert the image from BGR color (which OpenCV uses) to RGB color (which face_recognition uses)
-    rgb_small_frame = small_frame[:, :, ::-1]
-
-    # Only process every other frame of video to save time
-    if process_this_frame:
-        # Find all the faces and face encodings in the current frame of video
-        face_locations = face_recognition.face_locations(rgb_small_frame)
-        face_encodings = face_recognition.face_encodings(rgb_small_frame, face_locations)
-
-        face_names = []
-        for face_encoding in face_encodings:
-            # See if the face is a match for the known face(s)
-            matches = face_recognition.compare_faces(known_face_encodings, face_encoding)
-            name = "Unknown"
-
-            # If a match was found in known_face_encodings, just use the first one.
-            if True in matches:
-                first_match_index = matches.index(True)
-                name = known_face_names[first_match_index]
-
-            face_names.append(name)
-
-    process_this_frame = not process_this_frame
-
-
-    # Display the results
-    for (top, right, bottom, left), name in zip(face_locations, face_names):
-        # Scale back up face locations since the frame we detected in was scaled to 1/4 size
-        top *= 4
-        right *= 4
-        bottom *= 4
-        left *= 4
-
-        # Draw a box around the face
-        cv2.rectangle(frame, (left, top), (right, bottom), (0, 0, 255), 2)
-
-        # Draw a label with a name below the face
-        cv2.rectangle(frame, (left, bottom - 35), (right, bottom), (0, 0, 255), cv2.FILLED)
-        font = cv2.FONT_HERSHEY_DUPLEX
-        cv2.putText(frame, name, (left + 6, bottom - 6), font, 1.0, (255, 255, 255), 1)
-
-    # Display the resulting image
-    cv2.imshow('Video', frame)
-
-    # Hit 'q' on the keyboard to quit!
-    if cv2.waitKey(1) & 0xFF == ord('q'):
-        break
-
-# Release handle to the webcam
-video_capture.release()
-cv2.destroyAllWindows()
-
-
-
-#The process of face recognition given a frame of video, has the following basic steps (these all need to be decoupled from each other i.e. performed by different functions/classes):
-#1. Detect faces in the image
-#2. Align faces so they are all horizontal and properly zoomed
-#Then for each face:
-#3. Compute embeddings for the detected face
-#4. Compare this embedding to the embeddings for known faces in db (These were pre-computed when the face was first added
-
-#This whole process may require there to be a training directory where face images to be used for training will reside
-
-#An example implementation of this is what is above the description. A second example would be what is shown in https://www.pyimagesearch.com/2018/09/24/opencv-face-recognition/ . The second implementation has more nuts and bolts which means it is more flexible. It also allows you to do your own training so we could train with black faces as well. I suspect it will be more accurate but requires more work to implement.
