+++ requirements.txt
@@ -1,21 +1,24 @@
+amqp==5.0.6
 appdirs==1.4.3
 atomicwrites==1.4.0
 attrs==20.3.0
+billiard==3.6.4.0
 CacheControl==0.12.6
+celery==5.0.5
 certifi==2019.11.28
 chardet==3.0.4
 click==7.1.2
+click-didyoumean==0.0.3
+click-plugins==1.1.1
+click-repl==0.1.6
 colorama==0.4.3
 configparser==5.0.2
 contextlib2==0.6.0
 distlib==0.3.0
 distro==1.4.0
-dlib==19.22.0
 dnspython==2.1.0
 docopt==0.6.2
 email-validator==1.1.2
-face-recognition==1.3.0
-face-recognition-models==0.3.0
 Flask==1.1.2
 Flask-Login==0.5.0
 Flask-WTF==0.14.3
@@ -27,11 +30,13 @@ iniconfig==1.1.1
 ipaddr==2.2.0
 itsdangerous==1.1.0
 Jinja2==2.11.3
+kombu==5.0.2
 lockfile==0.12.2
 Markdown==3.3.4
 MarkupSafe==1.1.1
 more-itertools==8.7.0
 msgpack==0.6.2
+-e git+git@github.com:mcflyhalf/nipo.git@6a6adaec9962dc806fad950ad6f25784a01f3c67#egg=nipo
 numpy==1.20.2
 opencv-python==4.5.1.48
 packaging==20.3
@@ -41,6 +46,7 @@ pep517==0.8.2
 Pillow==8.2.0
 pluggy==0.13.1
 progress==1.5
+prompt-toolkit==3.0.18
 psycopg2-binary==2.8.6
 py==1.10.0
 Pygments==2.8.1
@@ -55,6 +61,8 @@ six==1.14.0
 SQLAlchemy==1.4.7
 toml==0.10.2
 urllib3==1.25.8
+vine==5.0.0
+wcwidth==0.2.5
 webencodings==0.5.1
 Werkzeug==1.0.1
 WTForms==2.3.3
